+++ Banmex/Banmex/Menu/ShowClients.cs
@@ -24,11 +24,14 @@ namespace Banmex.Menu
             accountdGridView.AutoSizeColumnsMode = DataGridViewAutoSizeColumnsMode.Fill;
         }
 
+        //metodo para cargar todos los datos a la tabla
         public void loadClientsData()
         {
             Connection.OpenConnection();
+            //se genera una lista con todos los clientes activos
             clientsGridView.DataSource = Class.Client.showAllClients(Connection.myConnection);
 
+            //se establecen los nombres de cabecera para las columnas
             this.clientsGridView.Columns[0].HeaderCell.Value = "ID";
             this.clientsGridView.Columns[1].HeaderCell.Value = "Nombre";
             this.clientsGridView.Columns[2].HeaderCell.Value = "Apellido";
@@ -40,11 +43,14 @@ namespace Banmex.Menu
             Connection.CloseConnection();
         }
 
+        //metodo para cargar todos los datos a la tabla
         public void loadAccountsData()
         {
             Connection.OpenConnection();
+            //se genera una lista con todas las cuentas activas de los clientes
             accountdGridView.DataSource = Class.Account.showAllAccounts(Connection.myConnection);
 
+            //se establecen los nombres de cabecera para las columnas
             this.accountdGridView.Columns[0].HeaderCell.Value = "ID Cuenta";
             this.accountdGridView.Columns[1].HeaderCell.Value = "ID Cliente";
             this.accountdGridView.Columns[2].HeaderCell.Value = "NIP";
@@ -57,6 +63,7 @@ namespace Banmex.Menu
             Connection.CloseConnection();
         }
 
+        //metodo de carga de las tablas cuado se abre la ventana
         private void ShowClients_Load(object sender, EventArgs e)
         {
             loadClientsData();
@@ -65,27 +72,47 @@ namespace Banmex.Menu
 
         private void modifyButton_Click(object sender, EventArgs e)
         {
-            string idClient = clientsGridView.CurrentRow.Cells[0].Value.ToString();
-            ModifyClient modWindow = new ModifyClient(Connection, idClient);
-            modWindow.ShowDialog();
-            loadClientsData();
-            loadAccountsData();
+            //validacion para la tabla clientes
+            //si la tabla esta vacia mostrar mensaje de error
+            if (clientsGridView.Rows.Count == 0)
+            {
+                MessageBox.Show("La tabla esta vacia");
+            }
+            else
+            {
+                //se toma el id del cliente
+                string idClient = clientsGridView.CurrentRow.Cells[0].Value.ToString();
+                //se manda al metodo modificar con el id
+                ModifyClient modWindow = new ModifyClient(Connection, idClient);
+                //se muestra la ventana con los datos
+                modWindow.ShowDialog();
+                //se recargan las tablas cuando regresa a esta ventana
+                loadClientsData();
+                loadAccountsData();
+            }
         }
 
         private void deleteButton_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("¿Seguro que desea eliminar este elemento?", "Confirmacion", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
-
-            if (result == DialogResult.Yes)
+            if(clientsGridView.Rows.Count == 0)
             {
-                string idClient = clientsGridView.CurrentRow.Cells[0].Value.ToString();
-                Connection.OpenConnection();
-                Class.Client.deleteClient(Connection.myConnection, idClient);
-                Connection.CloseConnection();
+                MessageBox.Show("La tabla esta vacia");
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("¿Seguro que desea eliminar este elemento?", "Confirmacion", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
 
-                MessageBox.Show("Eliminado exitosamente");
-                loadClientsData();
-                loadAccountsData();
+                if (result == DialogResult.Yes)
+                {
+                    string idClient = clientsGridView.CurrentRow.Cells[0].Value.ToString();
+                    Connection.OpenConnection();
+                    Class.Client.deleteClient(Connection.myConnection, idClient);
+                    Connection.CloseConnection();
+
+                    MessageBox.Show("Eliminado exitosamente");
+                    loadClientsData();
+                    loadAccountsData();
+                }
             }
         }
 
@@ -100,27 +127,45 @@ namespace Banmex.Menu
 
         private void modifyAccountButton_Click(object sender, EventArgs e)
         {
-            string idClient = accountdGridView.CurrentRow.Cells[0].Value.ToString();
-            ModifyAccount modWindow = new ModifyAccount(Connection, idClient);
-            modWindow.ShowDialog();
-            loadClientsData();
-            loadAccountsData();
+            //validacion para la tabla clientes
+            //si la tabla esta vacia mostrar mensaje de error
+            if (accountdGridView.Rows.Count == 0)
+            {
+                MessageBox.Show("La tabla esta vacia");
+            }
+            else
+            {
+                string idClient = accountdGridView.CurrentRow.Cells[0].Value.ToString();
+                ModifyAccount modWindow = new ModifyAccount(Connection, idClient);
+                modWindow.ShowDialog();
+                loadClientsData();
+                loadAccountsData();
+            }
         }
 
         private void deleteAccountButton_Click(object sender, EventArgs e)
         {
-            DialogResult result = MessageBox.Show("¿Seguro que desea eliminar este elemento?", "Confirmacion", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
-
-            if (result == DialogResult.Yes)
+            //validacion para la tabla clientes
+            //si la tabla esta vacia mostrar mensaje de error
+            if (accountdGridView.Rows.Count == 0)
             {
-                string idAccount = accountdGridView.CurrentRow.Cells[0].Value.ToString();
-                Connection.OpenConnection();
-                Class.Account.deleteAccount(Connection.myConnection, idAccount);
-                Connection.CloseConnection();
+                MessageBox.Show("La tabla esta vacia");
+            }
+            else
+            {
+                DialogResult result = MessageBox.Show("¿Seguro que desea eliminar este elemento?", "Confirmacion", MessageBoxButtons.YesNo, MessageBoxIcon.Information);
 
-                MessageBox.Show("Eliminado exitosamente");
-                loadClientsData();
-                loadAccountsData();
+                if (result == DialogResult.Yes)
+                {
+                    string idAccount = accountdGridView.CurrentRow.Cells[0].Value.ToString();
+                    Connection.OpenConnection();
+                    Class.Account.deleteAccount(Connection.myConnection, idAccount);
+                    Connection.CloseConnection();
+
+                    MessageBox.Show("Eliminado exitosamente");
+                    loadClientsData();
+                    loadAccountsData();
+                }
             }
         }
     }
